+++ Sources/SwiftSVG/PresentationAttributes.swift
@@ -2,14 +2,14 @@ import Swift2D
 
 public protocol PresentationAttributes {
     var fillColor: String? { get set }
-    var fillOpacity: Float? { get set }
+    var fillOpacity: Double? { get set }
     var fillRule: Fill.Rule? { get set }
     var strokeColor: String? { get set }
-    var strokeWidth: Float? { get set }
-    var strokeOpacity: Float? { get set }
+    var strokeWidth: Double? { get set }
+    var strokeOpacity: Double? { get set }
     var strokeLineCap: Stroke.LineCap? { get set }
     var strokeLineJoin: Stroke.LineJoin? { get set }
-    var strokeMiterLimit: Float? { get set }
+    var strokeMiterLimit: Double? { get set }
     var transform: String? { get set }
 }
