+++ Sources/SwiftSVG/Transformation.swift
@@ -32,9 +32,9 @@ import Swift2D
 /// | [W3](https://www.w3.org/TR/SVG11/coords.html#TransformAttribute)
 public enum Transformation {
     /// Moves an object by x & y. (Y is assumed to be '0' if not provided)
-    case translate(x: Float, y: Float)
+    case translate(x: Double, y: Double)
     /// Specifies a transformation in the form of a transformation matrix of six values.
-    case matrix(a: Float, b: Float, c: Float, d: Float, e: Float, f: Float)
+    case matrix(a: Double, b: Double, c: Double, d: Double, e: Double, f: Double)
     
     public enum Prefix: String, CaseIterable {
         case translate
@@ -64,7 +64,7 @@ public enum Transformation {
             var components = substring.split(separator: " ", omittingEmptySubsequences: true).map({ String($0) })
             components = components.flatMap({ $0.components(separatedBy: ",") })
             
-            let values = components.compactMap({ Float($0) }).map({ Float($0) })
+            let values = components.compactMap({ Double($0) }).map({ Double($0) })
             guard values.count > 0 else {
                 return nil
             }
@@ -90,7 +90,7 @@ public enum Transformation {
             var components = substring.split(separator: " ", omittingEmptySubsequences: true).map({ String($0) })
             components = components.flatMap({ $0.components(separatedBy: ",") })
             
-            let values = components.compactMap({ Float($0) }).map({ Float($0) })
+            let values = components.compactMap({ Double($0) }).map({ Double($0) })
             guard values.count > 5 else {
                 return nil
             }
